requirements.txt
streamlit
tensorflow
pillow
numpy
matplotlib
seaborn
pandas